Refactor data parsing for better list mapping performance
